added contentLength and Throughput calculations
By grabbing the content length of the file, and knowing the time from
request to download we can make estimates on the throughput of the
connection.

currently only used for reporting.. but you could imagine receiving
this data in analytics - and combined with radio network, signal
strength.. how a much more powerful algorithm for deciding network
throughput potentials could be made.